Fixed vendoring scripts and improved clarity.
No idea why I thought the `postinstall` thing would work correctly, I'm guessing I copied it from somewhere and didn't test it.  😬  There is now a `vendor` script in the root `package.json` which goes into the build directory, installs, and then runs the vendor script.  Doing an install every time is wasteful, but running the vendor script should be sufficiently uncommon that it isn't a big deal and having it "just work" with one command on first clone is nice.  We also do an `npm install` of the root in there just to make sure that the new dependencies have been actually picked up before vendoring them.  Again, this is an uncommon op so not a big deal on burning CPU/time.

The vendor script had an array of tuples that did not have much clarity on what each item was, so I changed it to an array of objects where the keys give some meaning to each entry, thus making it easier to modify.